refactor: adjust folder naming for profil procedures and separate profil io type file
